refactor and improve unit tests
